requirements.txt
flask==3.0.0
feedparser==6.0.11
google-generativeai==0.3.2
python-dotenv==1.0.0